requirements.txt
alabaster==0.7.12
attrs==18.2.0
autopep8==1.4.3
awscli==1.16.114
Babel==2.6.0
botocore==1.12.104
certifi==2018.11.29
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
colorama==0.3.9
coverage==4.5.2
docutils==0.14
entrypoints==0.3
flake8==3.7.7
haversine==2.0.0
idna==2.8
imagesize==1.1.0
Jinja2==2.10
jmespath==0.9.4
lxml==4.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.0.10
mkl-random==1.0.2
munch==2.3.2
numpy==1.16.2
packaging==19.0
pandas==0.24.1
pathlib-mate==0.0.15
psycopg2==2.7.7
pyasn1==0.4.5
pycodestyle==2.5.0
pyflakes==2.1.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.18
# Editable install with no version control (src==0.1.0)
-e /Users/sebastian/git_repos/data_science/scraper
urllib3==1.24.1
uszipcode==0.2.2
zipcode==4.0.0
